fix(admin): consolidate dashboard smoke into modeling-shell + fix 429
The standalone dashboard.spec.ts pushed cumulative login count past the
auth-rate-limiter (5/IP/15min) when run alongside multi-tenant-isolation,
producing 429s on subsequent specs. Folding the dashboard checks into
the existing single-login modeling-shell smoke keeps coverage without
extra logins.

Refs #356
